Variable logic optimization
